+++ packages/react-native-nitro-markdown/android/src/main/java/com/nitromarkdown/NitroMarkdownPackage.kt
@@ -5,11 +5,12 @@ import com.facebook.react.bridge.NativeModule
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.module.model.ReactModuleInfoProvider
 import com.facebook.react.uimanager.ViewManager
+import com.margelo.nitro.com.nitromarkdown.NitroMarkdownOnLoad
 
 class NitroMarkdownPackage : TurboReactPackage() {
     companion object {
         init {
-            System.loadLibrary("NitroMarkdown")
+            NitroMarkdownOnLoad.initializeNative()
         }
     }